This is a large rewrite of the core server. I eliminated hundreds of lines of code by pushing the socket level interactions into ebb.c. One would think this makes things unwieldy and difficult but on the contrary it makes the whole system much more comprehensible (albeit with some longer functions).
Requests are now allocated with the server. The only thing that is being
dynamically grown during a request is the GString buffer, but I would like to
eliminate that in the future too. On the Ruby side we're creating many Client
objects but this could be done away with by an immutable global array with
pre-initialized clients.

I'm amazed and happy at how the code has shrunk in this commit.