I optimized the ebb_client_env interface to allow for constant field names. There needs to be some more work to clean this up.
I also wrote a new benchmark to test how Ebb's speed was with respect to
response size. Unfortunately Ebb is running slower than evented mongrel for
small responses. More unfortunately this lead me to realize we cannot trust
previous benchmarks that I've made. The simple writing scheme I had in place
for development does not ensure the entire response gets written - it could,
and likely is the case, that many requests from ab were handled so quickly
because writing wasn't completeing! Oh how embarssing. But this can be fixed.
The next task is to use an real write buffer - any suggestions on which to
use?

Here is a graph. I believe x-axis is between 1 and 50 kilobytes response
size. The reason Ebb has a constant response speed is because it is
constantly writing only once to the socket.
http://chart.apis.google.com/chart?chxp=0,50&chdl=ebb|evented+mongrel|mongrel|thin&chtt=2008.01.27,requests+per+second,107+data+points&chd=s:BMYk9,bbbcc,BMYk9,1lZWP,BMYk9,UHEDC,BMYk9,9paWP&cht=lxy&chxt=x&chs=400x300&chxl=0:|kb&chco=444130,e4ac3d,f74343,7da478

There is much work to be done.