Improving POST body handling.
I added a cheap hack to pull the Content-Length out of the Mongrel parser for
use inside of Ebb. The hack is ugly and is only enough to get by before I can
figure out how to add a 'Content-Length' header to Ragel.

The next step is to change behavior so that Ebb reads the entire POST request
into memory (perhaps into a tmpfile) before giving it to Ruby. Ebb::Input will
be modified to provide a consistent interface to the tmpfile or memory.