Tests and a better FFI interface
The file benchmark/test.rb contains a handful of assertions to make very basic
demonstations that the server is doing what it intends to.

I've changed the FFI API so that it is more clear which functions are calling C
routines. Hopefully this will make ebb.rb a bit easier understand.

Large file uploads (greater than 16kb) are broken at the moment as is listening
from UNIX sockets.