Refactor POST interface
The C part of Ebb used to buffer the entire request (storing in a tempfile if
nessessary) before passing it off to Ruby. I've changed things around to allow
frameworks more control over uploads and also to simplify the C library.

Large file uploads still seem to be broken in the test case.