Add rb_thread_schedule() and threaded_processing option
This change should allow Ebb to work like mongrel with threaded processing or like Thin with sequential processing. There are trade off for each, thus i've made it a command line option '-S'

By default Ebb will use threaded processing because it seems that is working best. But only time will tell if that is the correction thing to do.