Use rb_thread_select instead of rb_thread_schedule
This cleans up the event loop processing greatly. I use rb_thread_select
to watch for changes on every file descriptor before calling ev_loop(loop,
EVLOOP_NONBLOCK); There is a 0.5 second timeout, to return to ruby for
checking @running and signals.

I still need to run benchmarks on this to make sure that it is still fast
:)