fixed nasty content-length bug.
apparently rails needs "CONTENT_LENGTH" key in the environ instead of "HTTP_CONTENT_LENGTH"?  the ragel parser is currently hitting content-length headers twice - this is a know bug and mearly an annoyance. I plan on working more with the parser soon so I'll leave the bug in for the 0.1.0 release.