Remove begin_transmission API
Instead set client->body_written = TRUE when ebb_client_release is called.
In side ebb_client_write_body() we will start the write watcher if it is not
already active and also write \r\n once all the headers are written. This
seems to clean up a lot of mess and confusion.