Major update to parser.rl
I'm changing Zed's original design of having a callback for each element in
the parser. Rather there will be two callbacks for headers: one for unknown
headers and one for known headers - there is an enum which defines the known
headers.

The idea here is to easily add preallocated global variables for often seen
header values in the ruby binding. This has only been done so far with
Content-Type, but I will add many more soon.

This change simplifies ebb.c greatly.